+++ lib/src/countly_config.dart
@@ -47,6 +47,7 @@ class CountlyConfig {
   String? _sdkBehaviorSettings;
   bool _backoffMechanismDisabled = false;
   bool _sdkBehaviorSettingsUpdatesDisabled = false;
+  int? _requestTimeoutDuration;
   bool _storingDefaultPushConsentDisabled = false;
   bool _viewRestartForManualRecordingDisabled = false;
 
@@ -145,6 +146,8 @@ class CountlyConfig {
 
   bool get sdkBehaviorSettingsUpdatesDisabled => _sdkBehaviorSettingsUpdatesDisabled;
 
+  int? get requestTimeoutDuration => _requestTimeoutDuration;
+  
   bool get storingDefaultPushConsentDisabled => _storingDefaultPushConsentDisabled;
 
   bool get viewRestartForManualRecordingDisabled => _viewRestartForManualRecordingDisabled;
@@ -408,6 +411,12 @@ class CountlyConfig {
     return this;
   }
 
+  /// Set the request timeout duration in seconds
+  /// [int requestTimeoutDuration] - duration in seconds
+  CountlyConfig setRequestTimeoutDuration(int requestTimeoutDuration) {
+    _requestTimeoutDuration = requestTimeoutDuration;
+  }
+  
   /// Disable storing the default push consent on initialization
   CountlyConfig disableStoringDefaultPushConsent() {
     _storingDefaultPushConsentDisabled = true;